feat: avoid printing empty line at end of output
If the delimiter is \n then we'll add two newlines at the end of the
message. This is good for printing multiple messages with the newline
delimiter so it's clear where each message begins and ends. However we
should avoid this for the very last line for a cleaner and more
consistent output. (all other instances should not have the empty new
line)